requirements.txt
alembic==0.9.1
appdirs==1.4.3
backports-abc==0.5
bleach==2.1.1
certifi==2017.7.27.1
click==6.7
configparser==3.5.0
dominate==2.3.1
enum34==1.1.6
et-xmlfile==1.0.1
flake8==3.3.0
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.2
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Moment==0.5.1
flask-mongoengine==0.9.3
Flask-PageDown==0.2.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
html5lib==1.0b10
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
livereload==2.5.1
lxml==4.1.1
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.14.3
numpy==1.13.3+mkl
olefile==0.44
openpyxl==2.4.9
packaging==16.8
Pillow==4.3.0
piltesseract==1.0.2
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
pymongo==3.5.1
pyparsing==2.2.0
python-docx==0.8.6
python-editor==1.0.3
scipy==1.0.0
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.2.0b3
tornado==4.5.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.12.1
WTForms==2.1
yapf==0.16.1
